+++ src/views/BrowseInventory/BrowseInventory.js
@@ -46,7 +46,10 @@ const BrowseInventory = () => {
   const { isFiltersOpened, toggleFilters } = useFiltersToogle(`${namespace}/filters`);
   const { deleteItemToView } = useItemToView('browse');
   const [pageConfig, setPageConfig] = useState(getLastBrowseOffset());
+  const hasFocusedSearchOptionOnMount = useRef(false);
   const inputRef = useRef();
+  const indexRef = useRef();
+  const paneTitleRef = useRef();
 
   const { search } = location;
 
@@ -81,13 +84,13 @@ const BrowseInventory = () => {
 
   const {
     data,
+    isFetched,
     isFetching,
     pagination,
     totalRecords,
   } = useInventoryBrowse({
     filters: withExtraFilters,
     pageParams: { pageConfig, setPageConfig },
-    options: { onSettled: deleteItemToView },
   });
 
   const { validateDataQuery } = useBrowseValidation(searchIndex);
@@ -128,7 +131,10 @@ const BrowseInventory = () => {
   const onApplySearch = useCallback(() => {
     const isSearchQueryValid = validateDataQuery(searchQuery);
 
-    if (isSearchQueryValid) applySearch();
+    if (isSearchQueryValid) {
+      applySearch();
+      paneTitleRef.current.focus();
+    }
   }, [searchQuery, filters]);
 
   const onChangeSearchIndex = useCallback((e) => {
@@ -140,6 +146,23 @@ const BrowseInventory = () => {
     resetFilters();
     inputRef.current.focus();
   }, [inputRef.current]);
+
+  useEffect(() => {
+    if (hasFocusedSearchOptionOnMount.current) {
+      return;
+    }
+
+    const handleSearchOptionFocus = () => {
+      indexRef.current.focus();
+      hasFocusedSearchOptionOnMount.current = true;
+    };
+
+    if (search && isFetched && !isFetching) {
+      handleSearchOptionFocus();
+    } else if (!search) {
+      handleSearchOptionFocus();
+    }
+  }, [isFetched, search, isFetching]);
 
   return (
     <PersistedPaneset
@@ -156,6 +179,7 @@ const BrowseInventory = () => {
           />
 
           <SingleSearchForm
+            autoFocus={false}
             applySearch={onApplySearch}
             changeSearch={changeSearch}
             disabled={!searchIndex}
@@ -167,6 +191,7 @@ const BrowseInventory = () => {
             selectedIndex={searchIndex}
             searchableIndexesPlaceholder={searchableIndexesPlaceholder}
             inputType="textarea"
+            indexRef={indexRef}
             inputRef={inputRef}
           />
 
@@ -189,6 +214,7 @@ const BrowseInventory = () => {
         isFetching={isFetching}
         isFiltersOpened={isFiltersOpened}
         pagination={pagination}
+        paneTitleRef={paneTitleRef}
         toggleFiltersPane={toggleFilters}
         totalRecords={totalRecords}
       />